Moved models to different project
