+++ conway.sln
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "conwaywebapi", "conwaywebapi\conwaywebapi.csproj", "{61BF698D-2757-4B60-AD27-FA8212B4188F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "conwaywebmodel", "conwaywebmodel\conwaywebmodel.csproj", "{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -74,6 +76,18 @@ Global
 		{61BF698D-2757-4B60-AD27-FA8212B4188F}.Release|x64.Build.0 = Release|Any CPU
 		{61BF698D-2757-4B60-AD27-FA8212B4188F}.Release|x86.ActiveCfg = Release|Any CPU
 		{61BF698D-2757-4B60-AD27-FA8212B4188F}.Release|x86.Build.0 = Release|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Debug|x64.Build.0 = Debug|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Debug|x86.Build.0 = Debug|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x64.ActiveCfg = Release|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x64.Build.0 = Release|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x86.ActiveCfg = Release|Any CPU
+		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
