+++ conwaywebapi/Database/Board.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ConwayWebApi.Models;
+using ConwayWebModel;
 
 namespace ConwayWebApi.Database
 {
@@ -23,8 +23,24 @@ namespace ConwayWebApi.Database
             Width = board.Info.Width;
             Height = board.Info.Height;
             Description = board.Info.Description;
-            
+
             BoardCells = board.LiveCells.Select(x => new BoardCell{X = x.X, Y = x.Y}).ToArray();
+        }
+        public BoardInfo ToBoardInfo()
+        {
+            return new BoardInfo{
+                Width = Width,
+                Height = Height,
+                Description = Description,
+                Id = ID
+            };
+        }
+        public BoardDetail ToBoardDetail()
+        {
+            return new BoardDetail{
+                Info = ToBoardInfo(),
+                LiveCells = BoardCells.Select(x => new CellCoord(x.X, x.Y)).ToArray()
+            };
         }
     }
 }
