+++ conwaywebapi/appsettings.json
@@ -1,7 +1,4 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "DefaultConnectionStringSecure"
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
