+++ conwaywebapi/conwaywebapi.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\conwaywebmodel\conwaywebmodel.csproj" />
+  </ItemGroup>
+
 </Project>
