+++ conwaywebmodel/BoardDetail.cs
@@ -1,8 +1,7 @@
 using System;
 using System.Linq;
-using ConwayWebApi.Database;
 
-namespace ConwayWebApi.Models
+namespace ConwayWebModel
 {
     public sealed class BoardDetail
     {
@@ -15,11 +14,6 @@ namespace ConwayWebApi.Models
         {
             Info = info;
             LiveCells = liveCells;
-        }
-        public BoardDetail(Board board, int? id)
-        {
-            Info = new BoardInfo(board, id);
-            LiveCells = board.BoardCells?.Select(x => new CellCoord(x.X, x.Y)).ToArray() ?? Array.Empty<CellCoord>();
         }
     }
 }
