+++ conwaywebmodel/BoardInfo.cs
@@ -1,6 +1,4 @@
-using ConwayWebApi.Database;
-
-namespace ConwayWebApi.Models
+namespace ConwayWebModel
 {
     public sealed class BoardInfo
     {
@@ -9,14 +7,6 @@ namespace ConwayWebApi.Models
             Width = width;
             Height = height;
             Description = description;
-            Id = id;
-        }
-        public BoardInfo(Board board, int? id = null)
-        {
-            
-            Width = board.Width;
-            Height = board.Height;
-            Description = board.Description;
             Id = id;
         }
         public BoardInfo(BoardInfo oldInfo, int? id = null)
